remove simple basis & refactor to three separate classes
