+++ gqcp/include/Basis/NonOrthogonalBasis/GNonOrthogonalStateBasis.hpp
@@ -19,87 +19,454 @@
 
 
 #include "Basis/BiorthogonalBasis/GLowdinPairingBasis.hpp"
-#include "Basis/NonOrthogonalBasis/SimpleNonOrthogonalStateBasis.hpp"
+#include "Basis/NonOrthogonalBasis/GNonOrthogonalStateBasis.hpp"
+#include "Basis/Transformations/BasisTransformable.hpp"
 #include "Basis/Transformations/GTransformation.hpp"
+#include "Basis/Transformations/JacobiRotatable.hpp"
+#include "Operator/FirstQuantized/NuclearRepulsionOperator.hpp"
 #include "Operator/SecondQuantized/GSQOneElectronOperator.hpp"
 #include "Operator/SecondQuantized/GSQTwoElectronOperator.hpp"
 #include "Operator/SecondQuantized/SQHamiltonian.hpp"
+#include "QuantumChemical/Spin.hpp"
+#include "QuantumChemical/SpinResolved.hpp"
+#include "Utilities/CRTP.hpp"
 
 
 namespace GQCP {
 
 
-/**
- *  MARK: GNonOrthogonalStateBasis implementation
+/*
+ *  MARK: GNonOrthogonalStateBasis
  */
 
-
 /**
- *  A type used to represent a non-orthogonal basis, generated from a set of generalized basis states.
+ *  A basis formed by any number of non-orthogonal states, in the form of `GTransformation`s.
  *
- *  @tparam _Scalar                 The scalar type used for the expansion coefficients of the states: real or complex.
+ *  @tparam _ExpansionScalar        The scalar type used to represent the expansion coefficients of the given non-orthogonal states: real or complex.
  */
 template <typename _Scalar>
 class GNonOrthogonalStateBasis:
-    public SimpleNonOrthogonalStateBasis<_Scalar, GNonOrthogonalStateBasis<_Scalar>> {
-public:
-    // The scalar type used for the expansion coefficients of the states: real or complex.
-    using Scalar = _Scalar;
+    public CRTP<GNonOrthogonalStateBasis<_Scalar>>,
+    public BasisTransformable<GNonOrthogonalStateBasis<_Scalar>>,
+    public JacobiRotatable<GNonOrthogonalStateBasis<_Scalar>> {
 
 public:
-    /**
+    // The scalar type used to represent the expansion coefficients of the given non-orthogonal states: real or complex.
+    using Scalar = _Scalar;
+
+    // The type of matrix associated with this kind of NonOrthogonalStateBasis.
+    using Matrix = SquareMatrix<Scalar>;
+
+    // The type of transformation that is naturally related to a `NonOrthogonalStateBasis`.
+    using Transformation = GTransformation<Scalar>;
+
+    // The type of non-orthogonal state basis this is.
+    using Self = GNonOrthogonalStateBasis<Scalar>;
+
+    // The biorthogonal basis related to the basis states in this type of non-orthogonal basis.
+    using BiorthogonalBasis = GLowdinPairingBasis<Scalar>;
+
+    // The second-quantized representation of any one-electron operator operator related to the `GNonOrthogonalStateBasis`.
+    using SQOverlapOperator = ScalarGSQOneElectronOperator<Scalar>;
+
+    // The second-quantized representation of any one-electron operator operator related to the `GNonOrthogonalStateBasis`.
+    using OneElectronOperator = ScalarGSQOneElectronOperator<Scalar>;
+
+    // The second-quantized representation of any two-electron operator operator related to the `GNonOrthogonalStateBasis`.
+    using TwoElectronOperator = ScalarGSQTwoElectronOperator<Scalar>;
+
+    // The second-quantized representation of the Hamiltonian that can be evaluated in this basis.
+    using Hamiltonian = GSQHamiltonian<Scalar>;
+
+    // The type of Jacobi rotation that is naturally related to the derived non orthogonal-basis.
+    using JacobiRotationType = typename JacobiRotatableTraits<Self>::JacobiRotationType;
+
+    // The vector containing the basis state of the associated type of transformations.
+    using States = std::vector<Transformation>;
+
+
+protected:
+    // The vector containing the non-orthogonal basis states.
+    States basis_states;
+
+    // The overlap operator in AO basis, constructed from the spinor basis.
+    SQOverlapOperator overlap_operator_AO;
+
+    // The total number of occupied orbitals.
+    size_t N;
+
+    // The threshold used to determine zero values.
+    double zero_threshold;
+
+
+public:
+    /*
      *  MARK: Constructors
      */
 
-    // Inherit `SimpleNonOrthogonalStateBasis`'s constructors.
-    using SimpleNonOrthogonalStateBasis<Scalar, GNonOrthogonalStateBasis<Scalar>>::SimpleNonOrthogonalStateBasis;
+    /**
+     *  Create a `GNonOrthogonalStateBasis` from any number of non orthogonal states.
+     *
+     *  @param basis_state_vector               The vector containing the non-orthogonal basis states.
+     *  @param S_AO                             The overlap operator in AO basis, constructed from the spinor basis used to calculate the non-orthogonal states.
+     *  @param number_of_occupied_orbitals      The total number of occupied orbitals in the system.
+     *  @param threshold                        The threshold at which a value is verified to be zero or not. The default is 1e-8.
+     */
+    GNonOrthogonalStateBasis<Scalar>(const States& basis_state_vector, const OneElectronOperator& S_AO, const size_t number_of_occupied_orbitals, const double threshold = 1e-8) :
+        basis_states {basis_state_vector},
+        overlap_operator_AO {S_AO},
+        N {number_of_occupied_orbitals},
+        zero_threshold {threshold} {
+
+        // The basis states must have the same dimensions.
+        for (size_t i = 0; i < basis_state_vector.size(); i++) {
+            if (basis_state_vector[0].dimension() != basis_state_vector[i].dimension()) {
+                throw std::invalid_argument("NonOrthogonalStateBasis<Scalar>(const States& basis_state_vector, const OneElectronOperator& S_AO, const size_t number_of_occupied_orbitals, const double threshold = 1e-8): The given basis states do not have the same dimensions.");
+            }
+        }
+    }
+
+
+    /*
+     *  MARK: Properties
+     */
+
+    /**
+     * Return the i'th basis states in the formed non-orthogonal state basis.
+     *
+     * @param i     The index of the basis state requested.
+     *
+     * @return The i'th basis state..
+     */
+    const Transformation& basisState(size_t i) const { return this->basis_states[i]; }
+
+    /**
+     * Return the dimension of the basis states in the formed non-orthogonal state basis.
+     *
+     * @return The dimension of the basis states.
+     *
+     * @note We return the dimension of the first state, as the constructor checks that all states have the same dimension.
+     */
+    const size_t basisStateDimension() const { return this->basis_states[0].dimension(); }
+
+    /**
+     * Return the basis states in the formed non-orthogonal state basis.
+     *
+     * @return The basis states.
+     */
+    const States& basisStates() const { return this->basis_states; }
+
+    /**
+     * Return the overlap metric of the formed non-orthogonal state basis as an operator.
+     *
+     * @return The overlap metric.
+     */
+    const OneElectronOperator& metric() const { return this->overlap_operator_AO; }
+
+    /**
+     * Return the number of basis states in the formed non-orthogonal state basis.
+     *
+     * @return The number of basis states.
+     */
+    const size_t numberOfBasisStates() const { return this->basis_states.size(); }
+
+    /**
+     * Return the number of occupied orbitals.
+     *
+     * @return The number of occupied orbitals.
+     */
+    const size_t numberOfOccupiedOrbitals() const { return this->N; }
+
+    /**
+     * Return the threshold used to compare values to zero associated with this non-orthogonal state basis.
+     *
+     * @return The threshold at which to evaluate zero values..
+     */
+    const double& threshold() const { return this->zero_threshold; }
+
+
+    /**
+     *  MARK: Conforming to `BasisTransformable`
+     */
+
+    /**
+     *  Apply the basis transformation and return the result.
+     *
+     *  @param T            The basis transformation.
+     *
+     *  @return The basis-transformed object.
+     */
+    Self transformed(const Transformation& T) const override {
+
+        auto result = this->derived();
+
+        for (size_t i = 0; i < result.numberOfBasisStates(); i++) {
+            result.basis_states[i].transform(T);
+        }
+
+        return result;
+    }
+
+    // Allow the `rotate` method from `BasisTransformable`, since there's also a `rotate` from `JacobiRotatable`.
+    using BasisTransformable<Self>::rotate;
+
+    // Allow the `rotated` method from `BasisTransformable`, since there's also a `rotated` from `JacobiRotatable`.
+    using BasisTransformable<Self>::rotated;
+
+
+    /**
+     *  MARK: Conforming to `JacobiRotatable`.
+     */
+
+    /**
+     *  Apply the Jacobi rotation and return the result.
+     *
+     *  @param jacobi_rotation          The Jacobi rotation.
+     *
+     *  @return The Jacobi-rotated object.
+     */
+    Self rotated(const JacobiRotationType& jacobi_rotation) const override {
+
+        const auto J = Transformation::FromJacobi(jacobi_rotation, this->basisStateDimension());
+        return this->rotated(J);
+    }
+
+    // Allow the `rotate` method from `JacobiRotatable`, since there's also a `rotate` from `BasisTransformable`.
+    using JacobiRotatable<Self>::rotate;
+
+
+    /**
+     * MARK: Operator evaluations
+     */
+
+    /**
+     * Evaluate any Hamiltonian operator in this non-orthogonal state basis.
+     *
+     * @param hamiltonian                       The Hamiltonian operator to be evaluated.
+     * @param nuclear_repulsion_operator        The nuclear repulsion operator associated with the molecule used for these calculations.
+     *
+     * @return The matrix representation of this Hamiltonian expressed in the non-orthogonal state basis.
+     *
+     * @note The given Hamiltonian must be expressed in the non-orthogonal AO basis.
+     * @note In order to express the Hamiltonian in this non-orthogonal state basis, we must take the nuclear repulsion into account as well (see equation (16) in the 2018 paper (https://doi.org/10.1063/1.4999218) by Olsen et. al. for example). This is why we also give the `nuclearRepulsionOperator` as a parameter.
+     */
+    Matrix evaluateHamiltonianOperator(const Hamiltonian& hamiltonian, const NuclearRepulsionOperator& nuclear_repulsion_operator) const {
+
+        // We first require the separate one and two electron operators from the Hamiltonian we want to evaluate.
+        const auto& h = hamiltonian.core();
+        const auto& g = hamiltonian.twoElectron();
+
+        // From the nuclear repulsion operator we need the value.
+        const auto nuc_rep = nuclear_repulsion_operator.value();
+
+        // Evaluate the operators in this basis.
+        // We also need the evaluated overlap in this basis.
+        const auto h_evaluated = this->evaluateOneElectronOperator(h);
+        const auto g_evaluated = this->evaluateTwoElectronOperator(g);
+        const auto s_evaluated = this->evaluateOverlapOperator();
+
+        // Return the Hamiltonian matrix, represented in this non-orthogonal state basis.
+        return h_evaluated + g_evaluated + (nuc_rep * s_evaluated);
+    }
+
+
+    /**
+     * Evaluate any scalar one-electron operator in this non-orthogonal state basis.
+     *
+     * @param f_op      The scalar one-electron operator to be evaluated.
+     *
+     * @return The matrix representation of this operator expressed in the non-orthogonal state basis.
+     *
+     * @note The given scalar one-electron operator must be expressed in the non-orthogonal AO basis.
+     * @note This implementation is based on equation 59 from the 2021 paper by Hugh Burton (https://aip.scitation.org/doi/abs/10.1063/5.0045442).
+     */
+    Matrix evaluateOneElectronOperator(const OneElectronOperator& f_op) const {
+
+        // We start by initializing a zero matrix of the correct dimension.
+        // Since the matrix is always squared, we only need one parameter to define its dimensions.
+        Matrix evaluated_operator = Matrix::Zero(this->numberOfBasisStates());
+
+        // Secondly, we map the operator parameters to a tensor representation.
+        // This will be necessary to perform the correct contractions.
+        Eigen::TensorMap<Eigen::Tensor<const Scalar, 2>> operator_tensor_map {f_op.parameters().data(), f_op.parameters().rows(), f_op.parameters().cols()};
+        Tensor<Scalar, 2> operator_parameters_tensor = Tensor<Scalar, 2>(operator_tensor_map);
+
+        // Loop over all basis states and calculate each matrix element using the generalized Slater-Condon rules.
+        for (size_t i = 0; i < this->numberOfBasisStates(); i++) {
+            for (size_t j = 0; j < this->numberOfBasisStates(); j++) {
+
+                // The first step is to create a biorthogonal basis from the two states that are being looped over.
+                const BiorthogonalBasis lowdin_pairing_basis {this->basisState(i), this->basisState(j), this->overlap_operator_AO, this->N};
+
+                // Check the number of zeros in the biorthogonal basis, as this influences how the matrix elements are calculated.
+                const auto number_of_zeros = lowdin_pairing_basis.numberOfZeroOverlaps();
+
+                // Use the correct formula, depending on how many zero overlaps there are.
+                // If there are zero overlap values, we perform the following calculation.
+                if (number_of_zeros == 0) {
+
+                    // In order to calcluate the matrix element when there are no zero overlap values, we need the reduced overlap and the weighted co-density matrix from the biorthogonal basis.
+                    const auto reduced_overlap = lowdin_pairing_basis.reducedOverlap();
+                    const auto weighted_co_density = lowdin_pairing_basis.weightedCoDensity();
+
+                    // Perform the contraction.
+                    Tensor<Scalar, 0> matrix_element = reduced_overlap * operator_parameters_tensor.template einsum<2>("uv, vu ->", weighted_co_density.matrix());
+                    evaluated_operator(i, j) += matrix_element(0);
+                }
+                // If there is one zero overlap value, we perform the following calculation.
+                else if (number_of_zeros == 1) {
+
+                    // In order to calcluate the matrix element when there are no zero overlap values, we need the reduced overlap.
+                    const auto reduced_overlap = lowdin_pairing_basis.reducedOverlap();
+
+                    // Next, calculate the co-density matrix of the orbital corresponding to the zero overlap value.
+                    // We know there is only one zero overlap orbital, so we acces the first index of the vector.
+                    const auto zero_overlap_index = lowdin_pairing_basis.zeroOverlapIndices()[0];
+                    const auto co_density = lowdin_pairing_basis.coDensity(zero_overlap_index);
+
+                    // Perform the contraction.
+                    Tensor<Scalar, 0> matrix_element = reduced_overlap * operator_parameters_tensor.template einsum<2>("uv, vu ->", co_density.matrix());
+                    evaluated_operator(i, j) += matrix_element(0);
+                }
+                // If there are two or more zero overlap values, the matrix element will be zero. No further if-clause is needed.
+            }
+        }
+
+        // Return the matrix representation of the evaluated operator.
+        return evaluated_operator;
+    }
+
+
+    /**
+     * Evaluate any overlap operator in this non-orthogonal state basis.
+     *
+     * @return The matrix representation of this overlap operator expressed in the non-orthogonal state basis.
+     *
+     * @note This method requires no parameters, since the overlap in AO basis is a member of the non-orthogonal state basis.
+     */
+    Matrix evaluateOverlapOperator() const {
+
+        // We start by initializing a zero matrix of the correct dimension.
+        // Since the matrix is always squared, we only need one parameter to define its dimensions.
+        Matrix evaluated_overlap = Matrix::Zero(this->numberOfBasisStates());
+
+        // Loop over all basis states and calculate the correct overlap elements using the biorthogonalized basis of each combination of states.
+        for (size_t i = 0; i < this->numberOfBasisStates(); i++) {
+            for (size_t j = 0; j < this->numberOfBasisStates(); j++) {
+
+                // The first step is to create a biorthogonal basis from the two states that are being looped over.
+                const BiorthogonalBasis lowdin_pairing_basis {this->basisState(i), this->basisState(j), this->overlap_operator_AO, this->N};
+
+                // Fill in the overlaps in the new matrix representation in this non-orthogonal basis.
+                evaluated_overlap(i, j) += lowdin_pairing_basis.totalOverlap();
+            }
+        }
+
+        // Return the overlap matrix.
+        return evaluated_overlap;
+    }
+
+
+    /**
+     * Evaluate any scalar two-electron operator in this non-orthogonal state basis.
+     *
+     * @param g_op      The scalar two-electron operator to be evaluated.
+     *
+     * @return The matrix representation of this operator expressed in the non-orthogonal state basis.
+     *
+     * @note The given scalar two-electron operator must be expressed in the non-orthogonal AO basis.
+     * @note This implementation is based on equation 65 from the 2021 paper by Hugh Burton (https://aip.scitation.org/doi/abs/10.1063/5.0045442).
+     */
+    Matrix evaluateTwoElectronOperator(const TwoElectronOperator& g_op) const {
+
+        // We start by initializing a zero matrix of the correct dimension.
+        // Since the matrix is always squared, we only need one parameter to define its dimensions.
+        Matrix evaluated_operator = Matrix::Zero(this->numberOfBasisStates());
+
+        // Loop over all basis states and calculate each matrix element using the generalized Slater-Condon rules.
+        for (size_t i = 0; i < this->numberOfBasisStates(); i++) {
+            for (size_t j = 0; j < this->numberOfBasisStates(); j++) {
+
+                // The first step is to create a biorthogonal basis from the two states that are being looped over.
+                const BiorthogonalBasis lowdin_pairing_basis {this->basisState(i), this->basisState(j), this->overlap_operator_AO, this->N};
+
+                // Check the number of zeros in the biorthogonal basis, as this influences how the matrix elements are calculated.
+                const auto number_of_zeros = lowdin_pairing_basis.numberOfZeroOverlaps();
+
+                // Use the correct formula, depending on how many zero overlaps there are.
+                // If there are zero overlap values, we perform the following calculation.
+                if (number_of_zeros == 0) {
+
+                    // In order to calcluate the matrix element when there are no zero overlap values, we need the reduced overlap and the weighted co-density matrix from the biorthogonal basis.
+                    const auto reduced_overlap = lowdin_pairing_basis.reducedOverlap();
+                    const auto weighted_co_density = lowdin_pairing_basis.weightedCoDensity();
+
+                    // Perform the contractions. We need to perform two contractions (one for the direct component, one for the exchange component).
+                    // In order to perserve readability of the contractions, and keep the exact link with the theory, we split each contraction in two.
+                    Tensor<Scalar, 2> intermediate_contraction_1 = g_op.parameters().template einsum<2>("utvs, tu -> vs", weighted_co_density.matrix());
+                    Tensor<Scalar, 0> direct_element = intermediate_contraction_1.template einsum<2>("vs, sv ->", weighted_co_density.matrix());
+
+                    Tensor<Scalar, 2> intermediate_contraction_2 = g_op.parameters().template einsum<2>("utvs, su -> tv", weighted_co_density.matrix());
+                    Tensor<Scalar, 0> exchange_element = intermediate_contraction_2.template einsum<2>("tv, tv ->", weighted_co_density.matrix());
+
+                    evaluated_operator(i, j) += (0.5 * reduced_overlap * (direct_element(0) - exchange_element(0)));
+                }
+                // If there is one zero overlap value, we perform the following calculation.
+                else if (number_of_zeros == 1) {
+
+                    // In order to calcluate the matrix element when there are no zero overlap values, we need the reduced overlap.
+                    const auto reduced_overlap = lowdin_pairing_basis.reducedOverlap();
+
+                    // Next, calculate the weighted co-density matrix and the co-density of the orbital corresponding to the zero overlap value.
+                    // We know there is only one zero overlap orbital, so we acces the first index of the vector.
+                    const auto zero_overlap_index = lowdin_pairing_basis.zeroOverlapIndices()[0];
+                    const auto co_density = lowdin_pairing_basis.coDensity(zero_overlap_index);
+                    const auto weighted_co_density = lowdin_pairing_basis.weightedCoDensity();
+
+                    // Perform the contractions. We need to perform two contractions (one for the direct component, one for the exchange component).
+                    // In order to perserve readability of the contractions, and keep the exact link with the theory, we split each contraction in two.
+                    Tensor<Scalar, 2> intermediate_contraction_1 = g_op.parameters().template einsum<2>("utvs, tu -> vs", co_density.matrix());
+                    Tensor<Scalar, 0> direct_element = intermediate_contraction_1.template einsum<2>("vs, sv ->", weighted_co_density.matrix());
+
+                    Tensor<Scalar, 2> intermediate_contraction_2 = g_op.parameters().template einsum<2>("utvs, su -> tv", co_density.matrix());
+                    Tensor<Scalar, 0> exchange_element = intermediate_contraction_2.template einsum<2>("tv, tv ->", weighted_co_density.matrix());
+
+                    evaluated_operator(i, j) += (reduced_overlap * (direct_element(0) - exchange_element(0)));
+                }
+                // If there are two zero overlap values, we perform the following calculation.
+                else if (number_of_zeros == 2) {
+
+                    // In order to calcluate the matrix element when there are no zero overlap values, we need the reduced overlap.
+                    const auto reduced_overlap = lowdin_pairing_basis.reducedOverlap();
+
+                    // Next, calculate the co-density of the orbitals corresponding to the zero overlap values.
+                    // We know there are only two zero overlap orbitals, so we acces the first and second index of the vector.
+                    const auto zero_overlap_index_1 = lowdin_pairing_basis.zeroOverlapIndices()[0];
+                    const auto zero_overlap_index_2 = lowdin_pairing_basis.zeroOverlapIndices()[1];
+                    const auto co_density_1 = lowdin_pairing_basis.coDensity(zero_overlap_index_1);
+                    const auto co_density_2 = lowdin_pairing_basis.coDensity(zero_overlap_index_2);
+
+                    // Perform the contractions. We need to perform two contractions (one for the direct component, one for the exchange component).
+                    // In order to perserve readability of the contractions, and keep the exact link with the theory, we split each contraction in two.
+                    Tensor<Scalar, 2> intermediate_contraction_1 = g_op.parameters().template einsum<2>("utvs, tu -> vs", co_density_1.matrix());
+                    Tensor<Scalar, 0> direct_element = intermediate_contraction_1.template einsum<2>("vs, sv ->", co_density_2.matrix());
+
+                    Tensor<Scalar, 2> intermediate_contraction_2 = g_op.parameters().template einsum<2>("utvs, su -> tv", co_density_1.matrix());
+                    Tensor<Scalar, 0> exchange_element = intermediate_contraction_2.template einsum<2>("tv, tv ->", co_density_2.matrix());
+
+                    evaluated_operator(i, j) += (reduced_overlap * (direct_element(0) - exchange_element(0)));
+                }
+                // If there are more than two zero overlap values, the matrix element will be zero. No further if-clause is needed.
+            }
+        }
+
+        // Return the matrix representation of the evaluated operator.
+        return evaluated_operator;
+    }
 };
 
-
-/*
- *  MARK: NonOrthogonalStateBasisTraits
- */
 
 /**
- *  A type that provides compile-time information on non-orthogonal state bases that is otherwise not accessible through a public class alias.
- */
-template <typename _Scalar>
-struct NonOrthogonalStateBasisTraits<GNonOrthogonalStateBasis<_Scalar>> {
-    // The scalar type used to represent a coefficient of the expansions of the states: real or complex.
-    using Scalar = _Scalar;
-
-    // The type of transformation that is naturally related to a `GNonOrthogonalStateBasis`.
-    using Transformation = GTransformation<Scalar>;
-
-    // The biorthogonal Löwdin pairing basis associated with the non-orthogonal state basis.
-    using BiorthogonalBasis = GLowdinPairingBasis<Scalar>;
-};
-
-
-/*
- *  MARK: NOSBasisOperatorTraits
- */
-
-/**
- *  A type that provides compile-time information on the operators associated with non-orthogonal state bases that is otherwise not accessible through a public class alias.
- */
-template <typename _Scalar>
-struct NOSBasisOperatorTraits<GNonOrthogonalStateBasis<_Scalar>> {
-    // The scalar type used to represent a coefficient of the expansions of the states: real or complex.
-    using Scalar = _Scalar;
-
-    // The second quantized representation of the Hamiltonian that is naturally related to a `GNonOrthogonalStateBasis`.
-    using Hamiltonian = GSQHamiltonian<Scalar>;
-
-    // The second-quantized representation of the one-electron operator that is naturally related to a `GNonOrthogonalStateBasis`.
-    using OneElectronOperator = ScalarGSQOneElectronOperator<Scalar>;
-
-    // The second-quantized representation of the two-electron operator that is naturally related to a `GNonOrthogonalStateBasis`.
-    using TwoElectronOperator = ScalarGSQTwoElectronOperator<Scalar>;
-};
-
-
-/*
  *  MARK: JacobiRotatableTraits
  */
 
@@ -114,7 +481,7 @@ struct JacobiRotatableTraits<GNonOrthogonalStateBasis<_Scalar>> {
 };
 
 
-/*
+/**
  *  MARK: BasisTransformableTraits
  */
 
@@ -124,7 +491,7 @@ struct JacobiRotatableTraits<GNonOrthogonalStateBasis<_Scalar>> {
 template <typename _Scalar>
 struct BasisTransformableTraits<GNonOrthogonalStateBasis<_Scalar>> {
 
-    // The type of transformation that is naturally related to a `GNonOrthogonalStateBasis`.
+    // The type of transformation that is naturally related to a `GNonOrthogonalStateBAsis`.
     using Transformation = GTransformation<_Scalar>;
 };
